Extension is end-to-end written but not functional
The extension is very similar to the rubocop-nova extension in terms of
architecture. Credo uses the exit status code to communicate the
category of the issue: https://hexdocs.pm/credo/exit_statuses.html

A future possibility is to make use of the status code. In any case a
non-zero status does not always indicate an error from Credo.