Extension works end-to-end
The extension successfully displays two credo issues in the text editor.

Adding a `return` to `toNovaIssue()` and fixing the mistakenly named
`novaIssues` in `provideIssues()` on line 185 fixed the extension. There
are also some minor corrections from `deno lint`. I removed `async` from
`inspect()` because I am unsure where a matching `await` should be or if
it is even needed.